requirements.txt
google-generativeai
speechrecognition
pyttsx3
python-dotenv